Introduce an explicit `copy var` notation (#247)
Taking into account a use-cases where a value class is used as an immutable handle into a mutable data-structure and the need to gradually teach developers a new features where properties of an immutable classes are somehow updated, we've decided to give an explicit name to the correspoding concept -- a "copyable property", and introduce a corresponding `copy` modifier both for "copyable properties" (`copy var`) and for copying functions (`copy fun`).

The text was updated to consistently use the corresponding terminology and the `copy` modififer. Motivation on the change was added into the text, too.